Rename graph_onnx -> onnx_graph
For consistency with pytorch_graph.py and caffe2_graph.py.
